+++ sdk/synapse/Azure.Analytics.Synapse.Artifacts/CHANGELOG.md
@@ -1,7 +1,7 @@
 # Release History
 
-## 1.0.0-beta.2 (Unreleased)
-
+## 1.0.0-preview.6 (Unreleased)
+- Expose some models
 
 ## 1.0.0-beta.1 (2020-11-12)
 - Supports List/Get Synapse resources through data plane APIs